segregate go1.8-only tests/benchs
