+++ controllers/main_test.go
@@ -1,9 +1,6 @@
 package controllers
 
 import (
-	mrand "math/rand"
-	"reflect"
-	"sort"
 	"testing"
 
 	"github.com/decred/dcrd/chaincfg"
@@ -27,112 +24,5 @@ func TestGetNetworkName(t *testing.T) {
 	if netName != "mainnet" {
 		t.Errorf("Incorrect network name: expected %s, got %s", "mainnet",
 			netName)
-	}
-}
-
-func randHashString() string {
-	var b [64]byte
-	const hexvals = "123456789abcdef"
-	for i := range b {
-		b[i] = hexvals[mrand.Intn(len(hexvals))]
-	}
-	return string(b[:])
-}
-
-func TestSortByTicketHeight(t *testing.T) {
-	// Create a large list of tickets to sort, voted over many blocks
-	ticketCount, maxTxHeight := 55000, int64(123000)
-
-	ticketInfoLive := make([]TicketInfoLive, 0, ticketCount)
-	for i := 0; i < ticketCount; i++ {
-		ticketInfoLive = append(ticketInfoLive, TicketInfoLive{
-			TicketHeight: uint32(mrand.Int63n(maxTxHeight)),
-			Ticket:       randHashString(), // could be nothing unless we sort with it
-		})
-	}
-
-	// Make a copy to sort with ref method
-	ticketInfoLive2 := make([]TicketInfoLive, len(ticketInfoLive))
-	copy(ticketInfoLive2, ticketInfoLive)
-
-	// Sort with ByTicketHeight, the test subject
-	sort.Sort(ByTicketHeight(ticketInfoLive))
-
-	// Sort using convenience function added in go1.8
-	sort.Slice(ticketInfoLive2, func(i, j int) bool {
-		return ticketInfoLive2[i].TicketHeight < ticketInfoLive2[j].TicketHeight
-	})
-	// compare
-	if !reflect.DeepEqual(ticketInfoLive, ticketInfoLive2) {
-		t.Error("Sort with ByTicketHeight failed")
-	}
-
-	// Check if sorted using convenience function added in go1.8
-	if !sort.SliceIsSorted(ticketInfoLive, func(i, j int) bool {
-		return ticketInfoLive[i].TicketHeight < ticketInfoLive[j].TicketHeight
-	}) {
-		t.Error("Sort with ByTicketHeight failed")
-	}
-}
-
-func BenchmarkSortByTicketHeight100(b *testing.B)   { benchmarkSortByTicketHeight(100, b) }
-func BenchmarkSortByTicketHeight500(b *testing.B)   { benchmarkSortByTicketHeight(500, b) }
-func BenchmarkSortByTicketHeight1000(b *testing.B)  { benchmarkSortByTicketHeight(1000, b) }
-func BenchmarkSortByTicketHeight2500(b *testing.B)  { benchmarkSortByTicketHeight(2500, b) }
-func BenchmarkSortByTicketHeight5000(b *testing.B)  { benchmarkSortByTicketHeight(5000, b) }
-func BenchmarkSortByTicketHeight10000(b *testing.B) { benchmarkSortByTicketHeight(10000, b) }
-func BenchmarkSortByTicketHeight20000(b *testing.B) { benchmarkSortByTicketHeight(20000, b) }
-
-func benchmarkSortByTicketHeight(ticketCount int, b *testing.B) {
-	// Create a large list of tickets to sort, voted over many blocks
-	maxTxHeight := int64(53000)
-
-	ticketInfoLive := make([]TicketInfoLive, 0, ticketCount)
-	for i := 0; i < ticketCount; i++ {
-		ticketInfoLive = append(ticketInfoLive, TicketInfoLive{
-			TicketHeight: uint32(mrand.Int63n(maxTxHeight)),
-			Ticket:       randHashString(), // could be nothing unless we sort with it
-		})
-	}
-
-	for i := 0; i < b.N; i++ {
-		// Make a copy to sort
-		ticketInfoLive2 := make([]TicketInfoLive, len(ticketInfoLive))
-		copy(ticketInfoLive2, ticketInfoLive)
-
-		// Sort with ByTicketHeight, the test subject
-		sort.Sort(ByTicketHeight(ticketInfoLive2))
-	}
-}
-
-func BenchmarkSortBySpentByHeight100(b *testing.B)   { benchmarkSortBySpentByHeight(100, b) }
-func BenchmarkSortBySpentByHeight500(b *testing.B)   { benchmarkSortBySpentByHeight(500, b) }
-func BenchmarkSortBySpentByHeight1000(b *testing.B)  { benchmarkSortBySpentByHeight(1000, b) }
-func BenchmarkSortBySpentByHeight2500(b *testing.B)  { benchmarkSortBySpentByHeight(2500, b) }
-func BenchmarkSortBySpentByHeight5000(b *testing.B)  { benchmarkSortBySpentByHeight(5000, b) }
-func BenchmarkSortBySpentByHeight10000(b *testing.B) { benchmarkSortBySpentByHeight(10000, b) }
-func BenchmarkSortBySpentByHeight20000(b *testing.B) { benchmarkSortBySpentByHeight(20000, b) }
-
-func benchmarkSortBySpentByHeight(ticketCount int, b *testing.B) {
-	// Create a large list of tickets to sort, voted over many blocks
-	maxTxHeight := int64(53000)
-
-	ticketInfoVoted := make([]TicketInfoHistoric, 0, ticketCount)
-	for i := 0; i < ticketCount; i++ {
-		ticketInfoVoted = append(ticketInfoVoted, TicketInfoHistoric{
-			Ticket:        randHashString(), // could be nothing unless we sort with it
-			SpentBy:       randHashString(),
-			SpentByHeight: uint32(mrand.Int63n(maxTxHeight)),
-			TicketHeight:  uint32(mrand.Int63n(maxTxHeight)),
-		})
-	}
-
-	for i := 0; i < b.N; i++ {
-		// Make a copy to sort
-		ticketInfoVoted2 := make([]TicketInfoHistoric, len(ticketInfoVoted))
-		copy(ticketInfoVoted2, ticketInfoVoted)
-
-		// Sort with BySpentByHeight, the test subject
-		sort.Sort(BySpentByHeight(ticketInfoVoted2))
 	}
 }
